refactor: replace contextx using gocommon

+++ internal/pkg/base/distance/distance.go
@@ -11,11 +11,11 @@ func Calculate(x1, y1, x2, y2 float64) float64 {
 func CalcWithGeo(lat1, lng1, lat2, lng2 float64, unit ...string) float64 {
 	const PI float64 = 3.141592653589793
 
-	radlat1 := float64(PI * lat1 / 180)
-	radlat2 := float64(PI * lat2 / 180)
+	radlat1 := PI * lat1 / 180
+	radlat2 := PI * lat2 / 180
 
-	theta := float64(lng1 - lng2)
-	radtheta := float64(PI * theta / 180)
+	theta := lng1 - lng2
+	radtheta := PI * theta / 180
 
 	dist := math.Sin(radlat1)*math.Sin(radlat2) + math.Cos(radlat1)*math.Cos(radlat2)*math.Cos(radtheta)
